Separated out Nvi navigation from basic data page
